+++ backend/server/openai/openai_response_formatter.py
